알람 감지

+++ app/src/main/java/com/example/universitybudgetapp/data/model/Entry.kt
@@ -7,9 +7,10 @@ import java.time.LocalDate
 @Entity(tableName = "entries")
 data class Entry(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val amount: Int,
+    val amount: Long,                  // Int → Long
     val description: String,
     val isIncome: Boolean,
-    val date: LocalDate,        // ✅ String → LocalDate
-    val category: Category      // ✅ CategoryConverter 이미 등록되어 있어야 함
+    val date: LocalDate,
+    val category: Category
 )
+
